Create Aseembly
The Atmega328p microcontroller for the arduino board has 32 internal 8-bit registers, R0-R31. R16-R31 can be used directly for i/o. The first instruction loads an 8-bit binary number into R16. The second instruction loads the value in R16 to the
DDRB register. Each bit of the DDRB register corresponds to a pin on the arduino.
The second instruction declares pin 5 to be an output port. Both the instructions
are equivalent to pinMode(5, OUTPUT)